// src/model/nfl/player.model.ts
export class Player {
    id: number;
    projectedPoints: number;
    fantasyPointsPerDollar: number;
    name: string;
    operator: string;
    opponent: string;
    opponentPositionRank: number;
    oponentRank: number;
    team: string;
    rank: number;
    position: string;
    salary: number;
    week: string;
}